+++ sdk/dataprotection/azure-mgmt-dataprotection/generated_samples/backup_instance_operations/put_backup_instance.py
@@ -82,6 +82,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: specification/dataprotection/resource-manager/Microsoft.DataProtection/stable/2023-05-01/examples/BackupInstanceOperations/PutBackupInstance.json
+# x-ms-original-file: specification/dataprotection/resource-manager/Microsoft.DataProtection/stable/2023-11-01/examples/BackupInstanceOperations/PutBackupInstance.json
 if __name__ == "__main__":
     main()